requirements.txt
ccxt
python-dotenv
rich
numpy
ta-lib
wheel